requirements.txt
notion-client==2.2.1
Mastodon.py==1.8.1
openai==1.12.0
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7

